# apps/desktop/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hashlark API",
    "description": "Search many torrent indexers at once. All routes except /health and /openapi.json need `Authorization: Bearer <token>`.",
    "contact": {
      "name": "Hashlark contributors"
    },
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/definitions/check": {
      "post": {
        "tags": [
          "definitions"
        ],
        "summary": "Validates a definition without saving it.",
        "operationId": "check_definition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefinitionCheck"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/history": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Recent searches, newest first.",
        "operationId": "get_history",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum entries (default 50, max 500).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntry"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "history"
        ],
        "summary": "Deletes all search history.",
        "operationId": "clear_history",
        "responses": {
          "204": {}
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/providers": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "Lists every provider with its state and health.",
        "operationId": "list_providers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderView"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Adds a provider. Re-adding a definition with the same id updates it.",
        "operationId": "add_provider",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProvider"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid definition; `details` lists the problems",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/providers/{id}": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "One provider.",
        "operationId": "get_provider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderView"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "providers"
        ],
        "summary": "Removes a user-added provider.",
        "operationId": "delete_provider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {},
          "400": {
            "description": "Built-in providers can't be removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "providers"
        ],
        "summary": "Changes a provider (enable/disable, rename, network policy).",
        "operationId": "update_provider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderView"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/resolve": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Gets the magnet link or `.torrent` URL for a result.",
        "operationId": "resolve",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadTarget"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The provider failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/results/{id}": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "A result from a recent search.",
        "operationId": "get_result",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Result id from a search",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergedResult"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Streams a search as Server-Sent Events.",
        "description": "Event names: `provider_started`, `results`, `provider_finished`,\n`provider_failed`, `done`. Each event's data is the JSON of a\n`SearchEvent`. `results` items with an `id` seen before replace it.\nClosing the connection cancels the search.",
        "operationId": "search_stream",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Text to search for.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cat",
            "in": "query",
            "description": "Comma-separated categories, e.g. `movies,tv`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "providers",
            "in": "query",
            "description": "Comma-separated provider ids. Default: all enabled providers.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imdb",
            "in": "query",
            "description": "IMDb id, for providers that support it.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "1-based page. Default 1.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "`relevance` (default), `title`, `seeders`, `peers`, `size` or `date`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of SearchEvent",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SearchEvent"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Runs a search to completion and returns sorted results.",
        "operationId": "search_blocking",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchOutcome"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Current settings.",
        "operationId": "get_settings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Replaces all settings. Fields left out take their defaults.",
        "operationId": "put_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Settings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddProvider": {
        "oneOf": [
          {
            "type": "object",
            "description": "From a YAML definition.",
            "required": [
              "yaml",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "definition"
                ]
              },
              "yaml": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "A Torznab endpoint (Jackett, Prowlarr, Bitmagnet).",
            "required": [
              "name",
              "url",
              "kind"
            ],
            "properties": {
              "api_key": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "kind": {
                "type": "string",
                "enum": [
                  "torznab"
                ]
              },
              "name": {
                "type": "string"
              },
              "url": {
                "type": "string"
              }
            }
          }
        ],
        "description": "How to add a provider."
      },
...
      "CheckDefinitionRequest": {
        "type": "object",
        "required": [
          "yaml"
        ],
        "properties": {
          "yaml": {
            "type": "string"
          }
        }
      },
...
      "DefinitionCheck": {
        "type": "object",
        "description": "Result of checking a definition without saving it.",
        "required": [
          "ok",
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "id": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "ok": {
            "type": "boolean"
          },
          "version": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "HistoryEntry": {
        "type": "object",
        "description": "One saved search.",
        "required": [
          "id",
          "query",
          "ts",
          "result_count"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "query": {
            "$ref": "#/components/schemas/SearchQuery"
          },
          "result_count": {
            "type": "integer",
            "format": "int64"
          },
          "ts": {
            "type": "integer",
            "format": "int64",
            "description": "Unix ms."
          }
        }
      },
...
      "ProviderView": {
        "type": "object",
        "description": "A provider as shown to users.",
        "required": [
          "id",
          "name",
          "kind",
          "description",
          "enabled",
          "builtin",
          "categories",
          "health",
          "network_policy",
          "source",
          "settings"
        ],
        "properties": {
          "builtin": {
            "type": "boolean",
            "description": "Shipped with Hashlark: can be disabled but not removed."
          },
          "capabilities": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Capabilities"
              },
              {
                "type": "null"
              }
            ]
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "description": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why the provider couldn't be loaded (e.g. an invalid definition)."
          },
          "health": {
            "$ref": "#/components/schemas/HealthSummary"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ProviderKind"
          },
          "name": {
            "type": "string"
          },
          "network_policy": {
            "$ref": "#/components/schemas/NetworkPolicy"
          },
          "settings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SettingView"
            }
          },
          "source": {
            "$ref": "#/components/schemas/ProviderSource"
          }
        }
      },
...
      "ResolveRequest": {
        "type": "object",
        "required": [
          "result_id"
        ],
        "properties": {
          "prefer": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TargetKind",
                "description": "Ask for a magnet or a `.torrent` when the result allows it."
              },
              {
                "type": "null"
              }
            ]
          },
          "result_id": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}